Fix EvID for MC (back to previously working version)

+++ PWGHF/treeHF/AliAnalysisTaskSEHFTreeCreator.h
@@ -235,7 +235,7 @@ public:
     void FillJetTree();
   
     
-    unsigned long GetEvID();
+    unsigned int GetEvID();
     
 private:
